+++ indico/core/celery/__init__.py
@@ -40,6 +40,7 @@ def _load_default_modules(app, **kwargs):
 
 @import_modules.connect
 def _import_modules(*args, **kwargs):
+    import indico.core.emails
     import indico.util.tasks
     signals.import_tasks.send()